+++ tests/integration/cases/set_mixed_int_widths/Scala_combined.scala
@@ -1,9 +1,9 @@
 object Fixture_set_mixed_int_widths_Scala_combined {
-var my_data = Set[Int](
+var my_data = Set[Long](
     1,
     1099511627776L,
 )
-my_data = Set[Int](
+my_data = Set[Long](
     1,
     1099511627776L,
 )
