+++ tests/integration/cases/set_mixed_int_widths/Scala_set_tree_set.scala
@@ -1,6 +1,6 @@
 import scala.collection.immutable.TreeSet
 object Fixture_set_mixed_int_widths_Scala_set_tree_set {
-val my_data = TreeSet[Int](
+val my_data = TreeSet[Long](
     1,
     1099511627776L,
 )
